refactor(codex): extract diagnostic-line parsing into codex.diagnostics.lib.mjs
src/codex.lib.mjs was 1367 lines, over the 1350-line warning threshold
(#1593). The codex_otel diagnostic parsing and the sub-session
reconstruction it feeds are a self-contained unit, so they move to
src/codex.diagnostics.lib.mjs unchanged.

codex.lib.mjs: 1367 -> 1250 lines.

Refs #2175
